+++ packages/soda-heroui/src/components/FormAutocomplete.tsx
@@ -7,15 +7,15 @@ import { FieldComponentProps } from "soda-tanstack-form"
 
 import { ErrorMessage } from "./ErrorMessage"
 
-export interface FormAutocompleteProps<FieldValue extends Key | undefined = Key | undefined, T extends object = object>
-    extends FieldComponentProps<typeof Autocomplete<T>, FieldValue> {}
+export interface FormAutocompleteProps<FieldValue extends Key | undefined = Key | undefined, RenderItem extends object = object>
+    extends FieldComponentProps<typeof Autocomplete<RenderItem>, FieldValue> {}
 
-export function FormAutocomplete<FieldValue extends Key | undefined = Key | undefined, T extends object = object>({
+export function FormAutocomplete<FieldValue extends Key | undefined = Key | undefined, RenderItem extends object = object>({
     field,
     ...rest
-}: FormAutocompleteProps<FieldValue, T>): ReactNode {
+}: FormAutocompleteProps<FieldValue, RenderItem>): ReactNode {
     return (
-        <Autocomplete<T>
+        <Autocomplete<RenderItem>
             selectedKey={field.state.value ?? null}
             onSelectionChange={value => field.handleChange((value ?? undefined) as FieldValue)}
             onBlur={field.handleBlur}